don't remove grade if it has classrooms,delete all classrooms at once , filter classrooms by grade

+++ app/Http/Controllers/GradeController.php
@@ -103,6 +103,9 @@ class GradeController extends Controller
     public function destroy(Grade $grade)
     {
         try{
+            if($grade->classrooms()){
+                return redirect()->back()->withError(__('messages.haveData',['first' => __('Classrooms_trans.grade') , 'second' => __('mainside.classrooms')]));
+            }
             $grade->delete();
             return redirect()->back()->withError(__('messages.delete'));
         }
